chore (refs T22614): move DpInlineNotification to client/js.
