+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/bundles/administrationMemberEmail.js
@@ -12,7 +12,7 @@
  */
 import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpEmailList from '@DemosPlanProcedureBundle/components/basicSettings/DpEmailList'
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 import { DpLabel } from 'demosplan-ui/components'
 import DpEditor from '@DpJs/components/core/DpEditor/DpEditor'
 import dpValidate from '@DpJs/lib/validation/dpValidate'
