+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/AdministrationImport/AdministrationImportNone.vue
@@ -8,7 +8,7 @@
 </license>
 
 <script>
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 
 export default {
   name: 'AdministrationImportNone',
