+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/AdministrationImport/EmailImport/EmailImport.vue
@@ -106,7 +106,7 @@
 import { CleanHtml } from 'demosplan-ui/directives'
 import { dpApi } from '@DemosPlanCoreBundle/plugins/DpApi'
 import DpDataTable from '@DpJs/components/core/DpDataTable/DpDataTable'
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 import { DpLoading } from 'demosplan-ui/components'
 import DpSimplifiedNewStatementForm from '@DemosPlanProcedureBundle/components/DpSimplifiedNewStatementForm'
 import EmailImportDate from './EmailImportDate'
