+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/AdministrationImport/ExcelImport/ExcelImportErrors.vue
@@ -74,7 +74,7 @@
 
 <script>
 import DpCheckbox from '@DpJs/components/core/form/DpCheckbox'
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 import DpProgressBar from '@DemosPlanCoreBundle/components/DpProgressBar'
 
 export default {
