+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/SegmentsBulkEdit/ActionStepper/ActionStepperResponse.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 
 export default {
   name: 'ActionStepperResponse',
