+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/admin/AdministrationPlaces.vue
@@ -124,7 +124,7 @@ import { dpApi, dpRpc } from '@DemosPlanCoreBundle/plugins/DpApi'
 import { DpButton, DpIcon, DpInput, DpLoading } from 'demosplan-ui/components'
 import DpButtonRow from '@DpJs/components/core/DpButtonRow'
 import DpDataTable from '@DpJs/components/core/DpDataTable/DpDataTable'
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 import dpValidateMixin from '@DpJs/lib/validation/dpValidateMixin'
 
 export default {
