+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/basicSettings/DpBasicSettings.vue
@@ -13,7 +13,7 @@ import { dpApi } from '@DemosPlanCoreBundle/plugins/DpApi'
 import DpDateRangePicker from '@DpJs/components/core/form/DpDateRangePicker'
 import DpDatetimePicker from '@DpJs/components/core/form/DpDatetimePicker'
 import DpEmailList from './DpEmailList'
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 import DpMultiselect from '@DpJs/components/core/form/DpMultiselect'
 import DpEditor from '@DpJs/components/core/DpEditor/DpEditor'
 import ExportSettings from './ExportSettings'
