+++ demosplan/DemosPlanStatementBundle/Resources/client/js/components/listStatements/ListStatements.vue
@@ -279,7 +279,7 @@ import DpBulkEditHeader from '@DemosPlanCoreBundle/components/DpBulkEditHeader'
 import DpClaim from '@DemosPlanStatementBundle/components/DpClaim'
 import DpDataTable from '@DpJs/components/core/DpDataTable/DpDataTable'
 import DpFlyout from '@DpJs/components/core/DpFlyout'
-import DpInlineNotification from '@DemosPlanCoreBundle/components/DpInlineNotification'
+import DpInlineNotification from '@DpJs/components/core/DpInlineNotification'
 import DpSelect from '@DpJs/components/core/form/DpSelect'
 import DpSlidingPagination from '@DemosPlanCoreBundle/components/DpSlidingPagination'
 import DpStickyElement from '@DpJs/components/core/shared/DpStickyElement'
